home message list: Changed narrow and message_lists to use HomeMessageList.
We now initialize the message_lists to use HomeMessageList for the home list.
We no longer specify the table name as this is set automatically by the class.
We update this.current when calling message_lists.set_current(), making sure
that any time the current page is changed, the HomeMessageList keeps track
of the state.

In narrow.js, we include handling the empty narrow banner inside of
narrow.deactivate, which is called when navigating to "All messages"
from a narrow. This is done because we don't always add messages
or rerender when navigating "All messages", so this now makes sure the banner
is updated.